Add query for getting gender identity options
