+++ VAMobile/src/api/types/DemographicsData.ts
@@ -11,3 +11,17 @@ export type UserDemographics = {
   genderIdentity: string
   preferredName: string
 }
+
+export type GenderIdentityOptions = {
+  [key: string]: string
+}
+
+export type GenderIdentityOptionsPayload = {
+  data: {
+    id: string
+    type: string
+    attributes: {
+      options: GenderIdentityOptions
+    }
+  }
+}
