+++ VAMobile/src/store/slices/personalInformationSlice.ts
@@ -5,8 +5,6 @@ import * as api from 'store/api'
 import {
   AddressData,
   AddressValidationScenarioTypes,
-  GenderIdentityOptions,
-  GenderIdentityOptionsData,
   PhoneData,
   PhoneType,
   ProfileFormattedFieldType,
@@ -59,8 +57,6 @@ export type PersonalInformationState = {
   showValidation: boolean
   preloadComplete: boolean
   validateAddressAbortController?: AbortController
-  genderIdentityOptions: GenderIdentityOptions
-  loadingGenderIdentityOptions: boolean
 }
 
 export const initialPersonalInformationState: PersonalInformationState = {
@@ -73,8 +69,6 @@ export const initialPersonalInformationState: PersonalInformationState = {
   preloadComplete: false,
   phoneNumberSaved: false,
   validateAddressAbortController: undefined,
-  genderIdentityOptions: {} as GenderIdentityOptions,
-  loadingGenderIdentityOptions: false,
 }
 
 const personalInformationNonFatalErrorString = 'Personal Information Service Error'
@@ -491,42 +485,6 @@ export const finishEditAddress = (): AppThunk => async (dispatch) => {
 }
 
 /**
- * Makes an API call to get valid gender identity options
- */
-export const getGenderIdentityOptions =
-  (screenID?: ScreenIDTypes): AppThunk =>
-  async (dispatch) => {
-    const retryFunction = () => dispatch(getGenderIdentityOptions(screenID))
-
-    try {
-      dispatch(dispatchClearErrors(screenID))
-      dispatch(dispatchSetTryAgainFunction(retryFunction))
-      dispatch(dispatchStartGetGenderIdentityOptions())
-
-      const response = await api.get<GenderIdentityOptionsData>('/v0/user/gender_identity/edit')
-      const responseOptions = response?.data.attributes.options || {}
-
-      // TODO: Look into adding an option to the API function for disabling the X-Key-Inflection property.
-      // Right now it's set to 'camel' which returns the keys in lowercase. We need to capitalize the keys
-      // so that they're consistent with what the PUT request for updating the genderIdentity field expects.
-      const genderIdentityOptions = Object.keys(responseOptions).reduce((options: GenderIdentityOptions, key: string) => {
-        options[key.toUpperCase()] = responseOptions[key]
-        return options
-      }, {})
-
-      dispatch(dispatchFinishGetGenderIdentityOptions({ genderIdentityOptions }))
-
-      await setAnalyticsUserProperty(UserAnalytics.vama_uses_profile())
-    } catch (error) {
-      if (isErrorObject(error)) {
-        logNonFatalErrorToFirebase(error, `getGenderIdentityOptions: ${personalInformationNonFatalErrorString}`)
-        dispatch(dispatchFinishGetGenderIdentityOptions({ error }))
-        dispatch(dispatchSetError({ errorType: getCommonErrorFromAPIError(error), screenID }))
-      }
-    }
-  }
-
-/**
  * Redux slice that will create the actions and reducers
  */
 const peronalInformationSlice = createSlice({
@@ -633,15 +591,6 @@ const peronalInformationSlice = createSlice({
       state.showValidation = !!addressData
       state.validateAddressAbortController = undefined
     },
-    dispatchStartGetGenderIdentityOptions: (state) => {
-      state.loadingGenderIdentityOptions = true
-    },
-    dispatchFinishGetGenderIdentityOptions: (state, action: PayloadAction<{ genderIdentityOptions?: GenderIdentityOptions; error?: Error }>) => {
-      const { genderIdentityOptions, error } = action.payload
-      state.loadingGenderIdentityOptions = false
-      state.genderIdentityOptions = genderIdentityOptions || {}
-      state.error = error
-    },
   },
 })
 
@@ -660,7 +609,5 @@ export const {
   dispatchStartSaveAddress,
   dispatchFinishValidateAddress,
   dispatchStartValidateAddress,
-  dispatchStartGetGenderIdentityOptions,
-  dispatchFinishGetGenderIdentityOptions,
 } = peronalInformationSlice.actions
 export default peronalInformationSlice.reducer
